requirements.txt
torch==2.0.1
torchvision==0.15.2
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch
ninja

mmdet==2.20.0 
mmengine==0.8.4 
mmsegmentation==0.20.0  
mmcls==0.25.0 
mmcv-full==1.5.0 -f https://download.openmmlab.com/mmcv/dist/cu118/torch2.0.1/index.html

pyyaml==6.0.1  
imageio==2.33.1 
imageio-ffmpeg==0.4.9 
lpips==0.1.4 
pytorch-msssim==1.0.0 
kornia==0.7.0 
yapf==0.40.1 
seaborn==0.13.2

crc32c 
pandas 
tensorboardX 
jupyterlab
matplotlib 
opencv-python 
scikit-image  
tqdm 
torchmetrics 